+++ examples/module-federation/base/remote/src/components/Button.tsx
@@ -1,0 +1,9 @@
+import './index.css';
+
+export const Button = () => {
+  return (
+    <button type="button" className="test">
+      Remote Button
+    </button>
+  );
+};