+++ packages/server/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/server-utils
 
+## 3.7.0
+
+### Patch Changes
+
+- @modern-js/utils@3.7.0
+
+## 3.6.0
+
+### Patch Changes
+
+- @modern-js/utils@3.6.0
+
 ## 3.5.0
 
 ### Patch Changes